Feat: Move test function to new module group. #103
